+++ src/Resources/contao/dca/tl_settings.php
@@ -3,7 +3,7 @@
 /**
  * palettes
  */
-$GLOBALS['TL_DCA']['tl_settings']['palettes']['default'] .= ';{counter_legend:hide},counter_topx_news,counter_topx_pages';
+$GLOBALS['TL_DCA']['tl_settings']['palettes']['default'] .= ';{counter_legend:hide},counter_topx_news,counter_topx_pages,counter_donotlog404,,counter_donotlogid';
 
 /**
  * Felder
@@ -35,4 +35,26 @@ $GLOBALS['TL_DCA']['tl_settings']['fields']['counter_topx_pages'] = array
 		'tl_class'            => 'w50'
 	),
 	'sql'                     => "int(10) NOT NULL default 100"
+);
+
+// Fehler 404 nicht im System-Log protokollieren
+$GLOBALS['TL_DCA']['tl_settings']['fields']['counter_donotlog404'] = array
+(
+	'label'                   => &$GLOBALS['TL_LANG']['tl_settings']['counter_donotlog404'],
+	'inputType'               => 'checkbox',
+	'eval'                    => array
+	(
+		'tl_class'            => 'w50 m12',
+	)
+);
+
+// Fehler source_id fehlend nicht im System-Log protokollieren
+$GLOBALS['TL_DCA']['tl_settings']['fields']['counter_donotlogid'] = array
+(
+	'label'                   => &$GLOBALS['TL_LANG']['tl_settings']['counter_donotlogid'],
+	'inputType'               => 'checkbox',
+	'eval'                    => array
+	(
+		'tl_class'            => 'w50 m12',
+	)
 );