+++ src/Resources/contao/languages/de/tl_settings.php
@@ -10,3 +10,5 @@ $GLOBALS['TL_LANG']['tl_settings']['counter_legend']       = 'Zähler';
  */
 $GLOBALS['TL_LANG']['tl_settings']['counter_topx_news'] = array('Anzahl Nachrichten', 'Anzahl der Nachrichten in der Top-Liste der Statistik');
 $GLOBALS['TL_LANG']['tl_settings']['counter_topx_pages'] = array('Anzahl Seiten', 'Anzahl der Seiten in der Top-Liste der Statistik');
+$GLOBALS['TL_LANG']['tl_settings']['counter_donotlog404'] = array('Fehler 404 nicht loggen', 'Fehler 404 nicht im System-Log protokollieren.');
+$GLOBALS['TL_LANG']['tl_settings']['counter_donotlogid'] = array('Fehlende source_id nicht loggen', 'Fehlende source_id nicht im System-Log protokollieren.');